Apply SoC
