+++ app/src/main/java/com/codesoom/assignment/controllers/TaskController.java
@@ -1,6 +1,7 @@
 package com.codesoom.assignment.controllers;
 
 import com.codesoom.assignment.NotFoundException;
+import com.codesoom.assignment.application.TaskService;
 import com.codesoom.assignment.models.Task;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -14,52 +15,41 @@ import java.util.Optional;
 @RequestMapping("/tasks")
 @CrossOrigin
 public class TaskController {
-    private List<Task> tasks = new ArrayList<>();
-    private Long newId = 0L;
+    private TaskService taskService;
+
+    public TaskController() {
+        taskService = new TaskService();
+    }
 
     @GetMapping
     public List<Task> list() {
-        return tasks;
+        return taskService.getTasks();
     }
 
     @GetMapping("/{id}")
     public Task get(@PathVariable Long id) throws IOException {
-        Task task = findById(id);
-        return task;
+        return taskService.getTask(id);
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public Task create(@RequestBody Task task) {
-        task.setId(generateId());
-        tasks.add(task);
-
-        return task;
+        return taskService.createTask(task);
     }
 
     @PutMapping("/{id}")
     public Task update(@PathVariable Long id, @RequestBody Task newTask) throws IOException {
-        Task oldTask = findById(id);
-        oldTask.setTitle(newTask.getTitle());
-        return newTask;
+        return taskService.updateTask(id, newTask);
+    }
+
+    @PatchMapping("/{id}")
+    public Task patch(@PathVariable Long id, @RequestBody Task newTask) throws IOException {
+        return taskService.updateTask(id, newTask);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) throws IOException {
-        Task task = findById(id);
-        tasks.remove(task);
-    }
-
-    private Long generateId() {
-        newId += 1;
-        return newId;
-    }
-
-    private Task findById(Long id) throws IOException {
-        return tasks.stream()
-                .filter(t -> t.getId().equals(id))
-                .findFirst()
-                .orElseThrow(NotFoundException::new);
+        taskService.deleteTask(id);
     }
 }
